feat: support multiple file input for CLI (#39)
* docs: organise docs for LLM

* docs: Add flatcitybuf specification, introduce new LLM and agent documentation, and refactor cursor rules.

* feat: Enable multi-file input with glob support and transform-aware merging for CityJSON serialization.

* feat: refactor CLI into a library, add e2e tests, and clean up benchmark artifact

* docs: update readme

+++ src/rust/fcb_core/src/reader/attr_query.rs
@@ -393,15 +393,14 @@ impl<R: Read + Seek> FcbReader<R> {
             magic_bytes: 8,
             header: header_size as u64,
             rtree_index: self.rtree_index_size(),
-            attributes: self.attr_index_size() as u64,
+            attributes: self.attr_index_size(),
         };
 
         let total_feat_count = result_vec.len() as u64;
 
         let attr_index_size = self.attr_index_size();
-        self.reader.seek(SeekFrom::Start(
-            attr_index_start_pos + attr_index_size as u64,
-        ))?;
+        self.reader
+            .seek(SeekFrom::Start(attr_index_start_pos + attr_index_size))?;
 
         Ok(FeatureIter::<R, Seekable>::new(
             self.reader,
@@ -475,7 +474,7 @@ impl<R: Read> FcbReader<R> {
             magic_bytes: 8,
             header: header_size as u64,
             rtree_index: self.rtree_index_size(),
-            attributes: self.attr_index_size() as u64,
+            attributes: self.attr_index_size(),
         };
 
         let total_feat_count = result.len() as u64;
